feat(pages): add WorkPlaceCard component to Tai Time page
This commit adds the WorkPlaceCard component to the Tai Time page, replacing the previous daily follow-up section. The WorkPlaceCard component will display information about a work place, such as the name of the site and the number of hours worked. This change enhances the user interface of the Tai Time page and makes it more user-friendly.